Added chunk-radiation for sure
this was so painful you have no idea. I've been working on this for a month. BUT NOW YOU CAN BE SHIELDED FROM RADIATION WOOOOO
(There is a possibility that it is still bugged, we'll find out when I implement a Geiger Counter)

Also added a test rad-resistant block

Next:

registering all damage types/sources and adding en_us for all of them.

Future:

Geiger Counter

registering all advancements and adding en_us for all of them.